containers/django-container/requirements.txt
Django>=3.2,<4.0
djangorestframework
psycopg2-binary
python-dotenv
pyspark==3.5.3